backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pymupdf==1.24.1
sentence-transformers==2.6.1
chromadb==0.5.3
rank-bm25==0.2.2
google-generativeai==0.7.2
